remove MaybeStar call from main.go
## Problem
when performing work, contributing to an open source project, or any other time the user is using `spr` they obviously don't want to be interrupted. It is worse than this interruption is for what is effectively spamming a meaningless metric.

## Solution

Remove the call that generates this spam. Future work can delete the code entirely.